+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.1.3 (unreleased)
 
 * updated import path to prevent download during docker builds
+* moved repository to newsdev
+* fixed an error that occurred when no keys were set for a group
+* fixed incorrect etcd address default for unset command
 
 ## 0.1.2